@@ -11,77 +11,96 @@
     "eas-android": "eas build --platform=android --auto-submit"
   },
   "dependencies": {
-    "@expo/vector-icons": "^12.0.0",
-    "@openspacelabs/react-native-zoomable-view": "^2.0.0",
-    "@react-native-async-storage/async-storage": "~1.15.0",
-    "@react-native-community/netinfo": "7.1.3",
-    "@react-native-picker/picker": "2.2.1",
-    "@react-navigation/bottom-tabs": "^6.0.9",
-    "@react-navigation/native": "^6.0.6",
-    "@react-navigation/stack": "^6.0.11",
-    "@reduxjs/toolkit": "^1.7.1",
-    "date-fns": "^2.28.0",
-    "expo": "^44.0.3",
-    "expo-app-loading": "~1.3.0",
-    "expo-application": "~4.0.1",
-    "expo-asset": "~8.4.6",
-    "expo-blur": "~11.0.0",
-    "expo-calendar": "~10.1.0",
-    "expo-camera": "~12.1.2",
-    "expo-clipboard": "~2.1.0",
-    "expo-device": "~4.1.0",
-    "expo-file-system": "~13.1.4",
-    "expo-font": "~10.0.4",
-    "expo-image-picker": "~12.0.1",
-    "expo-linking": "~3.0.0",
-    "expo-notifications": "~0.14.0",
-    "expo-random": "~12.1.1",
-    "expo-secure-store": "~11.1.0",
-    "expo-splash-screen": "~0.14.1",
-    "expo-status-bar": "~1.2.0",
-    "expo-web-browser": "~10.1.0",
-    "firebase": "^9.6.5",
-    "lodash-es": "^4.17.21",
+    "@native-base/icons": "^0.0.11",
+    "@openspacelabs/react-native-zoomable-view": "^2.0.4",
+    "@react-native-async-storage/async-storage": "^1.17.6",
+    "@react-native-community/netinfo": "8.2.0",
+    "@react-native-firebase/analytics": "^14.11.1",
+    "@react-native-firebase/app": "^14.11.1",
+    "@react-native-firebase/app-check": "^14.11.1",
+    "@react-native-firebase/auth": "^14.11.1",
+    "@react-native-firebase/crashlytics": "^14.11.1",
+    "@react-native-firebase/firestore": "^14.11.1",
+    "@react-native-firebase/functions": "^14.11.1",
+    "@react-native-firebase/remote-config": "^14.11.1",
+    "@react-native-firebase/storage": "^14.11.1",
+    "@react-native-picker/picker": "2.4.0",
+    "@react-navigation/bottom-tabs": "^6.3.1",
+    "@react-navigation/native": "^6.0.10",
+    "@react-navigation/stack": "^6.2.1",
+    "@reduxjs/toolkit": "^1.8.2",
+    "expo": "^45.0.0",
+    "expo-application": "~4.1.0",
+    "expo-asset": "~8.5.0",
+    "expo-auth-session": "~3.6.1",
+    "expo-blur": "~11.1.0",
+    "expo-calendar": "~10.2.0",
+    "expo-camera": "~12.2.0",
+    "expo-clipboard": "~3.0.1",
+    "expo-dev-client": "~1.0.0",
+    "expo-device": "~4.2.0",
+    "expo-file-system": "~14.0.0",
+    "expo-font": "~10.1.0",
+    "expo-image-picker": "~13.1.1",
+    "expo-linking": "~3.1.0",
+    "expo-notifications": "~0.15.4",
+    "expo-random": "~12.2.0",
+    "expo-secure-store": "~11.2.0",
+    "expo-splash-screen": "~0.15.1",
+    "expo-status-bar": "~1.3.0",
+    "expo-updates": "~0.13.2",
+    "expo-web-browser": "~10.2.1",
+    "lodash": "^4.17.21",
+    "luxon": "^2.4.0",
+    "lz-string": "^1.4.4",
+    "native-base": "^3.4.7",
     "node": "^16.13.0",
-    "react": "17.0.1",
-    "react-native": "0.64.3",
-    "react-native-elements": "^3.4.2",
-    "react-native-gesture-handler": "~2.1.0",
+    "react": "17.0.2",
+    "react-native": "0.68.2",
+    "react-native-gesture-handler": "~2.2.1",
     "react-native-open-maps": "^0.4.0",
-    "react-native-safe-area-context": "3.3.2",
-    "react-native-screens": "~3.10.1",
+    "react-native-safe-area-context": "4.2.4",
+    "react-native-screens": "~3.11.1",
     "react-native-super-grid": "^4.1.1",
-    "react-native-vector-icons": "^9.0.0",
-    "react-native-webview": "11.15.0",
-    "react-redux": "^7.2.6",
-    "redux": "^4.1.2",
-    "typescript": "~4.3.5"
+    "react-native-svg": "12.3.0",
+    "react-native-webview": "11.18.1",
+    "react-redux": "^8.0.2",
+    "redux": "^4.2.0",
+    "typescript": "^4.7.3",
+    "uuid": "3.4.0"
   },
   "devDependencies": {
     "@babel/core": "^7.12.9",
+    "@testing-library/jest-native": "^4.0.5",
+    "@testing-library/react-native": "^9.1.0",
     "@types/jest": "^27.5.0",
-    "@types/lodash-es": "^4.17.6",
-    "@types/react": "~17.0.21",
-    "@types/react-redux": "^7.1.22",
-    "@types/react-test-renderer": "^18.0.0",
+    "@types/lodash": "^4.14.182    ",
+    "@types/luxon": "^2.3.2",
+    "@types/lz-string": "^1.3.34",
+    "@types/react": "17.0.39",
+    "@types/react-native": "0.68.1",
+    "@types/react-redux": "^7.1.24",
     "@typescript-eslint/eslint-plugin": "^5.22.0",
     "@typescript-eslint/parser": "^5.22.0",
     "babel-jest": "^27.3.1",
     "eas-cli": "^0.46.0",
-    "eslint": "^8.14.0",
-    "eslint-config-prettier": "^8.5.0",
+    "eslint": "^8.18.0",
     "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-jest": "^26.5.3",
+    "eslint-plugin-jsdoc": "^39.3.2",
     "eslint-plugin-react": "^7.29.4",
+    "eslint-plugin-react-hooks": "^4.6.0",
     "eslint-plugin-react-native": "^4.0.0",
     "jest": "^26.6.3",
-    "jest-expo": "^44.0.0",
+    "jest-expo": "^45.0.0",
+    "metro-babel-register": "^0.71.1",
     "prettier": "^2.4.1",
-    "react-test-renderer": "^17.0.1",
-    "redux-devtools": "^3.7.0",
+    "react-native-bundle-visualizer": "^3.1.1",
     "ts-jest": "^26.0.0"
   },
   "resolutions": {
-    "react-devtools-core": "~4.14.0"
+    "react-devtools-core": "~4.14.0",
+    "@types/react": "~17.0.39"
   },
   "private": true
 }
