@@ -1,28 +1,19 @@
-import { View, ImageBackground, Image, StyleSheet } from "react-native";
-import { globalStyles } from "../../theme";
+import { Image, ImageBackground, View } from "react-native";
 
-const backgroundImg = require("../../../assets/home/db20_ribbon.jpg");
-const dbLogo = require("../../../assets/home/DB_Primary_Logo-01.png");
+
+import dbLogo from "../../../assets/home/DB_Primary_Logo-01.png";
+import backgroundImg from "../../../assets/home/db20_ribbon.jpg";
+import { globalStyles } from "../../theme";
 
 /**
  * A header image container used on the home screen
  */
 const HeaderImage = () => (
-  <View style={localStyles.headerContainer}>
+  <View>
     <ImageBackground source={backgroundImg} style={globalStyles.genericFillImage}>
-      <Image source={dbLogo} style={localStyles.dbLogo} />
+      <Image source={dbLogo} />
     </ImageBackground>
   </View>
 );
-
-const localStyles = {
-  headerContainer: StyleSheet.compose(globalStyles.genericCenteredView, {
-    height: 280,
-    width: "100%",
-    alignSelf: "center",
-    overflow: "hidden",
-  }),
-  dbLogo: StyleSheet.compose(globalStyles.genericIcon, { backgroundColor: "#FFFFFF99" }),
-};
 
 export default HeaderImage;
