@@ -1,12 +1,6 @@
 module.exports = {
   preset: "jest-expo",
-  globals: {
-    "ts-jest": {
-      tsconfig: {
-        jsx: "react-jsx",
-      },
-    },
-  },
+  globals: { "ts-jest": { tsconfig: { jsx: "react-jsx" } } },
   transform: {
     "^.+\\.tsx?$": "ts-jest",
     "^.+\\.js$": "<rootDir>/node_modules/react-native/jest/preprocessor.js",
@@ -19,9 +13,11 @@ module.exports = {
     "!**/babel.config.js",
     "!**/jest.setup.js",
   ],
-  moduleFileExtensions: ["js", "ts", "tsx"],
-  transformIgnorePatterns: [
-    "node_modules/(?!(jest-)?react-native|@react-native|react-clone-referenced-element|@react-native-community|expo(nent)?|@expo(nent)?/.*|react-navigation|@react-navigation/.*|@unimodules/.*|sentry-expo|native-base)",
+  moduleFileExtensions: [
+    "js", "ts", "tsx"
   ],
-  coverageReporters: ["json-summary", "text", "lcov"],
+  transformIgnorePatterns: ["node_modules/(?!(jest-)?react-native|@react-native|react-clone-referenced-element|@react-native-community|expo(nent)?|@expo(nent)?/.*|react-navigation|@react-navigation/.*|@unimodules/.*|sentry-expo|native-base)"],
+  coverageReporters: [
+    "json-summary", "text", "lcov"
+  ],
 };
