@@ -1,23 +1,24 @@
-import { getDownloadURL, ref } from "firebase/storage";
+import NetInfo, { NetInfoState, NetInfoStateType, NetInfoUnknownState } from "@react-native-community/netinfo";
+import firebaseStorage from "@react-native-firebase/storage";
+import { isEqual } from "lodash";
 import { DependencyList, useDebugValue, useEffect, useRef, useState } from "react";
-import { isEqual } from "lodash-es";
 import { TypedUseSelectorHook, useDispatch, useSelector } from "react-redux";
+
 import { AppDispatch, RootState } from "../redux/store";
-import { firebaseStorage } from "./FirebaseApp";
 
-export function useFirebaseStorageUrl(googleUri: string) {
-  useDebugValue(`Storage for ${googleUri}`);
+export function useFirebaseStorageUrl(googleUri?: string) {
+  useDebugValue(`Storage for ${googleUri ?? "undefined"}`);
 
-  const [state, setState] = useState<[string | null, Error | null]>([null, null]);
+  const [ state, setState ] = useState<[string | null, Error | null]>([ null, null ]);
 
   useEffect(() => {
     if (googleUri) {
-      getDownloadURL(ref(firebaseStorage, googleUri))
+      firebaseStorage().refFromURL(googleUri).getDownloadURL()
         .then((url) => {
-          setState([url, null]);
+          setState([ url, null ]);
         })
         .catch((error) => {
-          setState([null, error as Error]);
+          setState([ null, error as Error ]);
         });
     }
   }, [googleUri]);
@@ -38,18 +39,35 @@ export function useDeepEffect(effectFunc: () => unknown, deps: DependencyList) {
 
     isFirst.current = false;
     prevDeps.current = deps;
-  }, [deps, effectFunc]);
+  }, [ deps, effectFunc ]);
+}
+
+export function useNetworkStatus() {
+  const [ connectionInfo, setConnectionInfo ] = useState<NetInfoState>({
+    isConnected: null,
+    isInternetReachable: null,
+    type: NetInfoStateType.unknown,
+    details: null,
+  } as NetInfoUnknownState);
+  const [ isLoaded, setIsLoaded ] = useState(false);
+
+  useEffect(() => NetInfo.addEventListener((state) => {
+    setConnectionInfo(state);
+    setIsLoaded(true);
+  }));
+
+  return [ connectionInfo, isLoaded ] as const;
 }
 
 export function useCurrentDate(refreshInterval?: number) {
-  const [state, setState] = useState(new Date());
+  const [ state, setState ] = useState(new Date());
 
   useEffect(() => {
     // Set a *refreshInterval* second timer
     const timer = setInterval(() => {
       // Get time components
       setState(new Date());
-    }, (refreshInterval || 60) * 10);
+    }, (refreshInterval ?? 60) * 10);
 
     return () => {
       clearInterval(timer);
