@@ -6,7 +6,7 @@ import Badge from "./Badge";
 /**
  * A row of a user's {@link Badge}s loaded from Firebase
  */
-const Badges = ({ badges }) => (
+const Badges = ({ badges }: { badges: { name: string; image:string }[] }) => (
   <>
     {badges.length === 0 ? (
       <Text>You currently have no badges. Check back later to see if you&apos;ve earned any!</Text>
@@ -15,8 +15,8 @@ const Badges = ({ badges }) => (
         itemDimension={130}
         data={badges}
         spacing={10}
-        renderItem={({ item }) => (
-          <Badge name={item.name} imageURL={item.image} time={item.timeEarned} />
+        renderItem={({ item }: { item: { name: string; image:string } } ) => (
+          <Badge name={item.name} imageURL={item.image} />
         )}
       />
     )}
