@@ -1,168 +1,154 @@
+import { DateTime, Duration, Interval } from "luxon";
+import { Text } from "native-base";
 import { useEffect, useState } from "react";
-import { View, StyleSheet, ActivityIndicator } from "react-native";
-import { Text } from "react-native-elements";
-import { intervalToDuration } from "date-fns";
+import { StyleSheet, View } from "react-native";
 
-import { useAppSelector } from "../CustomHooks";
 import TimeUnit from "./TimeUnit";
 
-const CountdownView = () => {
-  const isConfigLoaded = false; //useAppSelector((state) => state.appConfig.isConfigLoaded);
-  const countdownConfig = { millis: 1 }; //useAppSelector((state) => state.appConfig.countdown);
-  const title = ""; // useAppSelector((state) => state.appConfig.countdown.title);
-  const [countdownDisplayDuration, setCountdownDisplayDuration] = useState<Duration>({
-    years: 0,
-    months: 0,
-    days: 0,
-    hours: 0,
-    minutes: 0,
-    seconds: 0,
-  });
+const CountdownView = ({
+  title, endTime
+}: { title: string; endTime: number }) => {
+  const [ countdownDisplayDuration, setCountdownDisplayDuration ] = useState<Duration>(Duration.fromMillis(0));
 
   useEffect(() => {
     // 1 second timer
     const timer = setInterval(() => {
       // Get time components
       setCountdownDisplayDuration(
-        intervalToDuration({ start: new Date(), end: new Date(countdownConfig.millis) })
+        Interval.fromDateTimes(new Date(), DateTime.fromMillis(endTime)).toDuration()
       );
     }, 1000);
 
     return () => {
       clearInterval(timer);
     };
-  }, [countdownConfig.millis]);
+  }, [endTime]);
 
   return (
     <View style={styles.container}>
-      {!isConfigLoaded && <ActivityIndicator color="white" size="large" />}
-      {isConfigLoaded && (
-        <>
-          <View style={styles.countdownTitleView}>
-            <Text adjustsFontSizeToFit numberOfLines={1} style={styles.countdownText}>
-              {title}
-            </Text>
-          </View>
-          <View style={styles.countdownTimer}>
-            {
-              // Check that this or any previous unit is nonzero
-              +!!countdownDisplayDuration.years > 0 && (
-                <>
-                  <TimeUnit unit="years" value={countdownDisplayDuration.years} />
+      <View style={styles.countdownTitleView}>
+        <Text adjustsFontSizeToFit numberOfLines={1}>
+          {title}
+        </Text>
+      </View>
+      <View style={styles.countdownTimer}>
+        {
+          // Check that this or any previous unit is nonzero
+          Number(!!countdownDisplayDuration.years) > 0 && (
+            <>
+              <TimeUnit unit="years" value={countdownDisplayDuration.years} />
 
-                  <Text
-                    style={{
-                      fontWeight: "bold",
-                      fontSize: 40,
-                      color: "white",
-                    }}
-                  >
+              <Text
+                style={{
+                  fontWeight: "bold",
+                  fontSize: 40,
+                  color: "white",
+                }}
+              >
                     :
-                  </Text>
-                </>
-              )
-            }
-            {
-              // Check that this or any previous unit is nonzero
-              +!!countdownDisplayDuration.years + +!!countdownDisplayDuration.months > 0 && (
-                <>
-                  <TimeUnit unit="months" value={countdownDisplayDuration.months} />
+              </Text>
+            </>
+          )
+        }
+        {
+          // Check that this or any previous unit is nonzero
+          Number(!!countdownDisplayDuration.years) + Number(!!countdownDisplayDuration.months) > 0 && (
+            <>
+              <TimeUnit unit="months" value={countdownDisplayDuration.months} />
 
-                  <Text
-                    style={{
-                      fontWeight: "bold",
-                      fontSize: 40,
-                      color: "white",
-                    }}
-                  >
+              <Text
+                style={{
+                  fontWeight: "bold",
+                  fontSize: 40,
+                  color: "white",
+                }}
+              >
                     :
-                  </Text>
-                </>
-              )
-            }
-            {
-              // Check that this or any previous unit is nonzero
-              +!!countdownDisplayDuration.years +
-                +!!countdownDisplayDuration.months +
-                +!!countdownDisplayDuration.days >
+              </Text>
+            </>
+          )
+        }
+        {
+          // Check that this or any previous unit is nonzero
+          Number(!!countdownDisplayDuration.years) +
+                Number(!!countdownDisplayDuration.months) +
+                Number(!!countdownDisplayDuration.days) >
                 0 && (
-                <>
-                  <TimeUnit unit="days" value={countdownDisplayDuration.days} />
+            <>
+              <TimeUnit unit="days" value={countdownDisplayDuration.days} />
 
-                  <Text
-                    style={{
-                      fontWeight: "bold",
-                      fontSize: 40,
-                      color: "white",
-                    }}
-                  >
+              <Text
+                style={{
+                  fontWeight: "bold",
+                  fontSize: 40,
+                  color: "white",
+                }}
+              >
                     :
-                  </Text>
-                </>
-              )
-            }
-            {
-              // Check that this or any previous unit is nonzero
-              +!!countdownDisplayDuration.years +
-                +!!countdownDisplayDuration.months +
-                +!!countdownDisplayDuration.days +
-                +!!countdownDisplayDuration.hours >
+              </Text>
+            </>
+          )
+        }
+        {
+          // Check that this or any previous unit is nonzero
+          Number(!!countdownDisplayDuration.years) +
+                Number(!!countdownDisplayDuration.months) +
+                Number(!!countdownDisplayDuration.days) +
+                Number(!!countdownDisplayDuration.hours) >
                 0 && (
-                <>
-                  <TimeUnit unit="hours" value={countdownDisplayDuration.hours} />
+            <>
+              <TimeUnit unit="hours" value={countdownDisplayDuration.hours} />
 
-                  <Text
-                    style={{
-                      fontWeight: "bold",
-                      fontSize: 40,
-                      color: "white",
-                    }}
-                  >
+              <Text
+                style={{
+                  fontWeight: "bold",
+                  fontSize: 40,
+                  color: "white",
+                }}
+              >
                     :
-                  </Text>
-                </>
-              )
-            }
-            {
-              // Check that this or any previous unit is nonzero
-              +!!countdownDisplayDuration.years +
-                +!!countdownDisplayDuration.months +
-                +!!countdownDisplayDuration.days +
-                +!!countdownDisplayDuration.hours +
-                +!!countdownDisplayDuration.minutes >
+              </Text>
+            </>
+          )
+        }
+        {
+          // Check that this or any previous unit is nonzero
+          Number(!!countdownDisplayDuration.years) +
+                Number(!!countdownDisplayDuration.months) +
+                Number(!!countdownDisplayDuration.days) +
+                Number(!!countdownDisplayDuration.hours) +
+                Number(!!countdownDisplayDuration.minutes) >
                 0 && (
-                <>
-                  <TimeUnit unit="minutes" value={countdownDisplayDuration.minutes} />
+            <>
+              <TimeUnit unit="minutes" value={countdownDisplayDuration.minutes} />
 
-                  <Text
-                    style={{
-                      fontWeight: "bold",
-                      fontSize: 40,
-                      color: "white",
-                    }}
-                  >
+              <Text
+                style={{
+                  fontWeight: "bold",
+                  fontSize: 40,
+                  color: "white",
+                }}
+              >
                     :
-                  </Text>
-                </>
-              )
-            }
-            {
-              // Check that this or any previous unit is nonzero
-              +!!countdownDisplayDuration.years +
-                +!!countdownDisplayDuration.months +
-                +!!countdownDisplayDuration.days +
-                +!!countdownDisplayDuration.hours +
-                +!!countdownDisplayDuration.minutes +
-                +!!countdownDisplayDuration.seconds >
+              </Text>
+            </>
+          )
+        }
+        {
+          // Check that this or any previous unit is nonzero
+          Number(!!countdownDisplayDuration.years) +
+                Number(!!countdownDisplayDuration.months) +
+                Number(!!countdownDisplayDuration.days) +
+                Number(!!countdownDisplayDuration.hours) +
+                Number(!!countdownDisplayDuration.minutes) +
+                Number(!!countdownDisplayDuration.seconds) >
                 0 && (
-                <>
-                  <TimeUnit unit="seconds" value={countdownDisplayDuration.seconds} />
-                </>
-              )
-            }
-          </View>
-        </>
-      )}
+            <>
+              <TimeUnit unit="seconds" value={countdownDisplayDuration.seconds} />
+            </>
+          )
+        }
+      </View>
     </View>
   );
 };
