@@ -1,13 +1,13 @@
 // From https://github.com/firebase/firebase-js-sdk/issues/97#issuecomment-452568889
 
-import { Platform, InteractionManager } from "react-native";
+import { InteractionManager, Platform } from "react-native";
 
 const _setTimeout = global.setTimeout;
 const _clearTimeout = global.clearTimeout;
 const MAX_TIMER_DURATION_MS = 60 * 1000;
 if (Platform.OS === "android") {
   // Work around issue `Setting a timer for long time`
-  // see: https://github.com/firebase/firebase-js-sdk/issues/97
+  // See: https://github.com/firebase/firebase-js-sdk/issues/97
   const timerFix = {};
   const runTask = (id, fn, ttl, args) => {
     const waitingTime = ttl - Date.now();
@@ -29,7 +29,7 @@ if (Platform.OS === "android") {
   global.setTimeout = (fn, time, ...args) => {
     if (MAX_TIMER_DURATION_MS < time) {
       const ttl = Date.now() + time;
-      const id = "_lt_" + Object.keys(timerFix).length;
+      const id = `_lt_${ Object.keys(timerFix).length}`;
       runTask(id, fn, ttl, args);
       return id;
     }
