@@ -1,39 +1,46 @@
+import firebaseRemoteConfig from "@react-native-firebase/remote-config";
 import { createAsyncThunk, createSlice } from "@reduxjs/toolkit";
-import { getAll } from "firebase/remote-config";
-import { showMessage } from "../common/AlertUtils";
-import { firebaseRemoteConfig } from "../common/FirebaseApp";
+
+import { showMessage } from "../common/util/AlertUtils";
+
 import { UserLoginType } from "./userDataSlice";
 
-type AppConfigSliceType = {
+interface AppConfigSliceType {
   isConfigLoaded: boolean;
   enabledScreens: string[];
-  countdowns: {
-    [key: string]: { millis: number; title: string };
-  };
   allowedLoginTypes: UserLoginType[];
+  scoreboardMode: { pointType:string;showIcons:boolean;showTrophies:boolean };
   demoModeKey: string;
-};
+}
 
 const initialState: AppConfigSliceType = {
   isConfigLoaded: false,
   enabledScreens: [],
-  countdowns: {},
   allowedLoginTypes: [],
+  scoreboardMode: { pointType: "", showIcons: false, showTrophies: false },
   demoModeKey: Math.random().toString(), // Start the demo key as junk for safety's sake
 };
 
 export const updateConfig = createAsyncThunk(
   "appConfig/updateConfig",
   async (): Promise<Partial<AppConfigSliceType>> => {
-    const remoteConfig = getAll(firebaseRemoteConfig);
-    console.log(remoteConfig);
+    const remoteConfigInstance = firebaseRemoteConfig();
+    await remoteConfigInstance.fetchAndActivate();
+    const remoteConfig = {} as Partial<AppConfigSliceType>;
+
+    remoteConfig.enabledScreens = (JSON.parse(remoteConfigInstance.getString("rn_shown_tabs")) ?? undefined) as string[] | undefined;
+    remoteConfig.allowedLoginTypes = ((JSON.parse(remoteConfigInstance.getString("login_mode")) as { rn?: UserLoginType[] }).rn ?? undefined);
+    remoteConfig.demoModeKey = remoteConfigInstance.getString("demo_mode_key");
+    remoteConfig.scoreboardMode = (JSON.parse(remoteConfigInstance.getString("rn_scoreboard_mode")) ?? undefined) as {
+      pointType: string;
+      showIcons: boolean;
+      showTrophies: boolean;
+  } | undefined;
+
+    return remoteConfig;
   }
 );
 
-// Redux Toolkit allows us to write "mutating" logic in reducers. It
-// doesn't actually mutate the state because it uses the Immer library,
-// which detects changes to a "draft state" and produces a brand new
-// immutable state based off those changes
 export const appConfigSlice = createSlice({
   name: "appConfig",
   initialState,
@@ -46,7 +53,7 @@ export const appConfigSlice = createSlice({
       Object.assign(state, {
         isConfigLoaded: false,
         enabledScreens: [],
-        countdowns: {},
+        scoreboardMode: {},
         allowedLoginTypes: [],
         demoModeKey: Math.random().toString(), // Start the demo key as junk for safety's sake
       });
@@ -59,15 +66,15 @@ export const appConfigSlice = createSlice({
         Object.assign(state, initialState);
       })
       .addCase(updateConfig.fulfilled, (state, action) => {
-        state.enabledScreens = action.payload.enabledScreens || initialState.enabledScreens;
-        state.countdowns = action.payload.countdowns || initialState.countdowns;
-        state.allowedLoginTypes = action.payload.allowedLoginTypes || initialState.allowedLoginTypes;
-        state.demoModeKey = action.payload.demoModeKey || initialState.demoModeKey;
+        state.enabledScreens = action.payload.enabledScreens ?? initialState.enabledScreens;
+        state.scoreboardMode = action.payload.scoreboardMode ?? initialState.scoreboardMode;
+        state.allowedLoginTypes = action.payload.allowedLoginTypes ?? initialState.allowedLoginTypes;
+        state.demoModeKey = action.payload.demoModeKey ?? initialState.demoModeKey;
         state.isConfigLoaded = true;
       })
       .addCase(updateConfig.rejected, (state, action) => {
         state.isConfigLoaded = true;
-        showMessage(action.error.message ?? '', action.error.code, undefined, true, action.error.stack);
+        showMessage(action.error.message ?? "", action.error.code, undefined, true, action.error.stack);
       });
   },
 });
