@@ -1,14 +1,22 @@
 import { configureStore } from "@reduxjs/toolkit";
 
-import authReducer from "./authSlice";
-import notificationReducer from "./notificationSlice";
 import appConfigReducer from "./appConfigSlice";
+import authReducer from "./authSlice";
+import deviceDataReducer from "./deviceDataSlice";
+import globalLoadingReducer from "./globalLoadingSlice";
+import marathonReducer from "./marathonSlice";
+import notificationReducer from "./notificationSlice";
+import userDataReducer from "./userDataSlice";
 
 const store = configureStore({
   reducer: {
-    auth: authReducer,
-    notification: notificationReducer,
     appConfig: appConfigReducer,
+    auth: authReducer,
+    deviceData: deviceDataReducer,
+    globalLoading: globalLoadingReducer,
+    marathon: marathonReducer,
+    notification: notificationReducer,
+    userData: userDataReducer,
   },
 });
 export default store;
