@@ -1,10 +1,10 @@
 // Allow importing from the parent folder rather than individual files
-// for example you can do:   *import { EventScreen } from '../EventScreen'*
-// instead of:               *import { EventScreen } from '../EventScreen/EventScreen'*
+// For example you can do:   *import { EventScreen } from '../EventScreen'*
+// Instead of:               *import { EventScreen } from '../EventScreen/EventScreen'*
 
-import EventView from "./EventView";
 import EventRow from "./EventRow";
 import EventScreen from "./EventScreen";
+import EventView from "./EventView";
 
 export { EventRow, EventView };
 export default EventScreen;
