@@ -1,6 +1,6 @@
 // Allow importing from the parent folder rather than individual files
-// for example you can do:   *import { EventScreen } from '../EventScreen'*
-// instead of:               *import { EventScreen } from '../EventScreen/EventScreen'*
+// For example you can do:   *import { EventScreen } from '../EventScreen'*
+// Instead of:               *import { EventScreen } from '../EventScreen/EventScreen'*
 
 import HeaderImage from "./HeaderImage";
 import HomeScreen from "./HomeScreen";
