@@ -1,29 +1,22 @@
-import {
-  addDoc,
-  arrayRemove,
-  arrayUnion,
-  collection,
-  doc,
-  onSnapshot,
-  setDoc,
-  updateDoc,
-} from "firebase/firestore";
-import { ref, uploadBytes } from "firebase/storage";
-import * as ImagePicker from "expo-image-picker";
-import { useEffect, useState } from "react";
+/* eslint-disable */
+import firebaseAuth from "@react-native-firebase/auth";
+import firebaseFirestore from "@react-native-firebase/firestore";
+import firebaseStorage from "@react-native-firebase/storage";
 import { Picker } from "@react-native-picker/picker";
+import * as ImagePicker from "expo-image-picker";
+import { createRef, useEffect, useState } from "react";
 import { ActionSheetIOS, Platform, Text, TextInput, View } from "react-native";
-import { Button, Divider, Input } from "react-native-elements";
-import { showMessage } from "../../common/AlertUtils";
-import { firebaseAuth, firebaseFirestore, firebaseStorage } from "../../common/FirebaseApp";
-import store from "../../redux/store";
-import generateUuid from "../../common/GenerateUuid";
-import { globalColors, globalTextStyles } from "../../theme";
+import { Button, Divider, Input } from "native-base";
+
 import Standings from "../../common/components/Standings";
+import { showMessage } from "../../common/util/AlertUtils";
+import generateUuid from "../../common/util/GenerateUuid";
+import store from "../../redux/store";
+import { globalColors, globalTextStyles } from "../../theme";
 import { StandingType } from "../../types/StandingType";
 
-const guessInput = React.createRef<TextInput>();
-const dadInput = React.createRef<TextInput>();
+const guessInput = createRef<TextInput>();
+const dadInput = createRef<TextInput>();
 
 // From https://github.com/expo/examples/blob/master/with-firebase-storage-upload/App.js
 async function uploadImageAsync(uri: string, hour: string) {
@@ -50,38 +43,38 @@ async function uploadImageAsync(uri: string, hour: string) {
     }
   );
 
-  const fileRef = ref(
-    firebaseStorage,
+  const fileRef = firebaseStorage().refFromURL(
     `gs://react-danceblue.appspot.com/marathon/2022/photo-booth/${generateUuid()}`
   );
-  const result = await uploadBytes(fileRef, blob);
+  const result = await fileRef.put(blob);
   success = success && !!result.ref;
 
   // We're done with the blob, close and release it
   blob.close();
 
   if (success) {
-    addDoc(collection(firebaseFirestore, "marathon/2022/photo-booth"), {
+    firebaseFirestore().collection("marathon/2022/photo-booth").add({
       photoUri: result.ref.fullPath,
       hour,
-      name: `${store.getState().auth.firstName?.toString()} ${store
+      name: `${store.getState().userData.firstName?.toString()} ${store
         .getState()
-        .auth.lastName?.toString()}`,
-      linkblue: store.getState().auth.linkblue || null,
-      email: store.getState().auth.email || null,
-      moraleTeamID: store.getState().auth.moraleTeamId || null,
+        .userData.lastName?.toString()}`,
+      linkblue: store.getState().userData.linkblue || null,
+      email: store.getState().userData.email || null,
+      // moraleTeamID: store.getState().userData.moraleTeamId || null,
       deviceId: store.getState().notification.uuid || null,
       pushToken: store.getState().notification.pushToken || null,
-    }).then(() => {
-      showMessage("Photo uploaded", "Success");
-    });
+    })
+      .then(() => {
+        showMessage("Photo uploaded", "Success");
+      });
   }
 
   return success;
 }
 
 // TODO move this elsewhere
-async function getImageFromPhotosOrCamera(
+function getImageFromPhotosOrCamera(
   callback: (uri: string, hour: string) => Promise<boolean>,
   hour: string
 ) {
@@ -110,7 +103,7 @@ async function getImageFromPhotosOrCamera(
 
     const pickerResult = await ImagePicker.launchImageLibraryAsync({
       allowsEditing: true,
-      aspect: [4, 3],
+      aspect: [ 4, 3 ],
     });
 
     try {
@@ -150,7 +143,7 @@ async function getImageFromPhotosOrCamera(
 
     const pickerResult = await ImagePicker.launchCameraAsync({
       allowsEditing: true,
-      aspect: [4, 3],
+      aspect: [ 4, 3 ],
     });
 
     if (!(await takePictureCallback((pickerResult as ImagePicker.ImageInfo).uri, hour))) {
@@ -158,14 +151,12 @@ async function getImageFromPhotosOrCamera(
     }
   };
 
-  if (store.getState().appConfig.offline) {
-    showMessage(
-      "You seem to be offline, connect to the internet or talk to your morale group leader"
-    );
-  } else if (Platform.OS === "ios") {
+  if (Platform.OS === "ios") {
     ActionSheetIOS.showActionSheetWithOptions(
       {
-        options: ["Cancel", "Choose from Photos", "Take a Photo"],
+        options: [
+          "Cancel", "Choose from Photos", "Take a Photo"
+        ],
         cancelButtonIndex: 0,
       },
       (buttonIndex) => {
@@ -186,8 +177,8 @@ async function getImageFromPhotosOrCamera(
 
 // Life is a Highway
 export const PhotoUpload = () => {
-  const [uploading, setUploading] = useState(false);
-  const [selectedStation, setSelectedStation] = useState("Colorado");
+  const [ uploading, setUploading ] = useState(false);
+  const [ selectedStation, setSelectedStation ] = useState("Colorado");
 
   return (
     <View
@@ -224,9 +215,13 @@ export const PhotoUpload = () => {
         disabled={uploading}
         onPress={() => {
           setUploading(true);
-          getImageFromPhotosOrCamera(uploadImageAsync, selectedStation).then(() =>
-            setUploading(false)
-          );
+          getImageFromPhotosOrCamera(
+            (uri: string, hour: string) => {
+              return uploadImageAsync(uri, hour).then((val) => {
+                setUploading(false);
+                return val;
+              });
+            }, selectedStation);
         }}
       />
     </View>
@@ -235,25 +230,22 @@ export const PhotoUpload = () => {
 
 // Dad Joke
 export const DadJokeLeaderboard = () => {
-  const dadJokesDoc = doc(firebaseFirestore, "marathon/2022/dad-jokes/dad-jokes");
-  const [dadJokes, setDadJokes] = useState<StandingType[]>([]);
+  const dadJokesDoc = firebaseFirestore().doc("marathon/2022/dad-jokes/dad-jokes");
+  const [ dadJokes, setDadJokes ] = useState<StandingType[]>([]);
 
-  useEffect(() =>
-    onSnapshot(dadJokesDoc, (dadJokeSnapshot) => {
-      const dadJokeSnapshotData = dadJokeSnapshot.data() as {
-        [key: string]: { jokeText: string; votes: string[] };
-      };
-      if (dadJokeSnapshotData) {
-        const dadJokeSnapshotArray = Object.entries(dadJokeSnapshotData);
-        const tempDadJokes = dadJokeSnapshotArray.map(([id, joke]) => ({
-          id,
-          name: joke.jokeText,
-          points: joke.votes.length,
-          highlighted: joke.votes.indexOf(firebaseAuth.currentUser?.uid || "") > -1, // Use highlighted to mark if it should be start checked
-        }));
-        setDadJokes(tempDadJokes);
-      }
-    })
+  useEffect(() => dadJokesDoc.onSnapshot((dadJokeSnapshot) => {
+    const dadJokeSnapshotData = dadJokeSnapshot.data() as Record<string, { jokeText: string; votes: string[] }>;
+    if (dadJokeSnapshotData) {
+      const dadJokeSnapshotArray = Object.entries(dadJokeSnapshotData);
+      const tempDadJokes = dadJokeSnapshotArray.map(([ id, joke ]) => ({
+        id,
+        name: joke.jokeText,
+        points: joke.votes.length,
+        highlighted: joke.votes.includes(firebaseAuth().currentUser?.uid || ""), // Use highlighted to mark if it should be start checked
+      }));
+      setDadJokes(tempDadJokes);
+    }
+  })
   );
 
   return (
@@ -269,45 +261,37 @@ export const DadJokeLeaderboard = () => {
         standingData={dadJokes}
         expandable
         collapsedRows={5}
-        dadJokeTempMagic
-        dadJokeTempMagicCallback={(checked, id) => {
-          if (firebaseAuth.currentUser?.uid) {
-            if (checked) {
-              updateDoc(dadJokesDoc, `${id}.votes`, arrayUnion(firebaseAuth.currentUser.uid));
-            } else {
-              updateDoc(dadJokesDoc, `${id}.votes`, arrayRemove(firebaseAuth.currentUser.uid));
-            }
-          }
-        }}
       />
       <Input
         ref={dadInput}
         disabled={
-          !firebaseAuth.currentUser ||
-          !firebaseAuth.currentUser.uid ||
-          firebaseAuth.currentUser.isAnonymous
+          !firebaseAuth().currentUser ||
+          !firebaseAuth().currentUser?.uid ||
+          firebaseAuth().currentUser?.isAnonymous
         }
         defaultValue={
-          !firebaseAuth.currentUser || firebaseAuth.currentUser.isAnonymous
+          !firebaseAuth().currentUser || firebaseAuth().currentUser?.isAnonymous
             ? " LinkBlue login required"
             : " Enter your dad joke and press enter"
         }
         style={{ borderColor: "blue", borderWidth: 1, borderRadius: 5, marginTop: 10 }}
         autoCompleteType="off"
         autoComplete="off"
-        onSubmitEditing={async (event) => {
+        onSubmitEditing={(event) => {
           const submittedText = event.nativeEvent.text;
 
-          if (firebaseAuth.currentUser?.uid) {
-            await updateDoc(
-              doc(firebaseFirestore, "marathon/2022/dad-jokes/dad-jokes"),
-              firebaseAuth.currentUser.uid,
+          if (firebaseAuth().currentUser?.uid) {
+            firebaseFirestore().doc("marathon/2022/dad-jokes/dad-jokes").update(
+              {
+                [firebaseAuth().currentUser?.uid ?? ""]:
               { jokeText: submittedText, votes: [] }
-            ).then(() => {
-              if (dadInput.current) {
-                dadInput.current.clear();
               }
-            });
+            )
+              .then(() => {
+                if (dadInput.current) {
+                  dadInput.current.clear();
+                }
+              });
           }
         }}
       />
@@ -340,20 +324,17 @@ const activities = {
             !Number.isNaN(parsedText) &&
             parsedText > 0
           ) {
-            await setDoc(
-              doc(
-                collection(firebaseFirestore, "marathon/2022/guessing-game"),
-                firebaseAuth.currentUser?.uid
-              ),
+            await firebaseFirestore().doc(`marathon/2022/guessing-game/${firebaseAuth().currentUser?.uid ?? ""}`).set(
               {
                 guess: parsedText,
-                email: firebaseAuth.currentUser?.email,
+                email: firebaseAuth().currentUser?.email,
               }
-            ).then(() => {
-              if (guessInput.current) {
-                guessInput.current.clear();
-              }
-            });
+            )
+              .then(() => {
+                if (guessInput.current) {
+                  guessInput.current.clear();
+                }
+              });
           } else {
             showMessage("Make sure to enter a positive number", "Invalid Guess", () => {
               if (guessInput.current) {
@@ -384,6 +365,6 @@ const activities = {
       onPress={() => getImageFromPhotosOrCamera(uploadImageAsync, "photo-booth")}
     />
   ),
-} as { [key: string]: JSX.Element };
+} as Record<string, JSX.Element>;
 
 export default activities;
