@@ -1,21 +1,20 @@
 import { useNavigation } from "@react-navigation/native";
-import { differenceInHours } from "date-fns";
-import { useEffect, useMemo, useState } from "react";
-import { FlatList, Platform, StyleSheet, useWindowDimensions, View } from "react-native";
-import { Divider, Image, ListItem, Text } from "react-native-elements";
-import CountdownView from "../../common/components/CountdownView";
+import { DateTime, Interval } from "luxon";
+import { Button, Divider, Image, Text } from "native-base";
+import { useEffect, useState } from "react";
+import { FlatList, View, useWindowDimensions } from "react-native";
+
 import { useCachedFiles } from "../../common/CacheUtils";
 import { useAppSelector, useCurrentDate } from "../../common/CustomHooks";
-import { globalColors } from "../../theme";
-import { FirestoreHour } from "../../types/FirebaseTypes";
-import { TabScreenProps } from "../../types/NavigationTypes";
-import store from "../../redux/store";
 import { appConfigSlice, updateConfig } from "../../redux/appConfigSlice";
+import store from "../../redux/store";
+import { FirestoreHour } from "../../types/FirebaseTypes";
+import { TabNavigatorProps } from "../../types/NavigationTypes";
 
 function revealRandomChars(input: string, charsToReveal: number): string {
   let tempOutputString = "";
-  for (let i = 0; i < input.length; i++) {
-    if (input[i] === " ") {
+  for (const char of input) {
+    if (char === " ") {
       tempOutputString += " ";
     } else {
       tempOutputString += "■";
@@ -30,8 +29,7 @@ function revealRandomChars(input: string, charsToReveal: number): string {
 
   for (let i = 0; i < charsToReveal; i++) {
     const inputIndex = Math.floor(input.length * ((Math.sin(seedNumber * i) + 1) / 2));
-    tempOutputString =
-      tempOutputString.substring(0, inputIndex) +
+    tempOutputString = tempOutputString.substring(0, inputIndex) +
       input[inputIndex] +
       tempOutputString.substring(inputIndex + 1);
   }
@@ -50,12 +48,14 @@ const HourRow = ({
   marathonHour: number;
   currentMinute: number;
 }) => {
-  const { name: hourName, hourNumber } = firestoreHour;
-  const navigation = useNavigation<TabScreenProps<"HoursScreen">["navigation"]>();
-  const [displayedName, setDisplayedName] = useState("");
-  const [clickable, setClickable] = useState(false);
+  const {
+    name: hourName, hourNumber
+  } = firestoreHour;
+  const navigation = useNavigation<TabNavigatorProps<"Marathon">["navigation"]>();
+  const [ displayedName, setDisplayedName ] = useState("");
+  const [ clickable, setClickable ] = useState(false);
 
-  // TODO change this so ti looks like wordle and reveals random letters up to half of the name; whole name at the hour so reveal stays fresh
+  // TODO change this so it looks like wordle and reveals random letters up to half of the name; whole name at the hour so reveal stays fresh
   // Maybe choose which to reveal based on hash of name? Need to be the same every time
   useEffect(() => {
     let tempDisplayedName = "";
@@ -63,19 +63,17 @@ const HourRow = ({
     if (marathonHour + 1 > hourNumber) {
       tempDisplayedName = hourName;
       setClickable(true);
-    }
-    // Should we check if we should start revealing this one?
-    else {
+    } else {
+      // Should we check if we should start revealing this one?
       setClickable(false);
       if (marathonHour === hourNumber - 1) {
         const hourPercent = (currentMinute + 1) / 60;
         const percentNameToShow = (hourPercent - 0.75) * 4;
-        const charsToShow =
-          percentNameToShow > 0 ? Math.trunc(hourName.length * percentNameToShow) : 0;
+        const charsToShow = percentNameToShow > 0 ? Math.trunc(hourName.length * percentNameToShow) : 0;
         tempDisplayedName = revealRandomChars(hourName, charsToShow);
       } else {
-        for (let i = 0; i < hourName.length; i++) {
-          if (hourName[i] === " ") {
+        for (const char of hourName) {
+          if (char === " ") {
             tempDisplayedName += " ";
           } else {
             tempDisplayedName += "■";
@@ -84,36 +82,35 @@ const HourRow = ({
       }
     }
     setDisplayedName(tempDisplayedName);
-  }, [currentMinute, hourName, hourNumber, marathonHour]);
+  }, [
+    currentMinute, hourName, hourNumber, marathonHour
+  ]);
 
   return (
-    <ListItem
-      hasTVPreferredFocus={undefined}
-      tvParallaxProperties={undefined}
+    <Button
       onPress={
-        clickable ? () => navigation?.navigate("Hour Details", { firestoreHour }) : undefined
+        clickable ? () => navigation.navigate("Hour Details", { firestoreHour }) : undefined
       }
       disabled={!clickable}
       key={hourNumber}
     >
-      <ListItem.Content style={{ flexDirection: "row", justifyContent: "flex-start" }}>
-        <Text h4 h4Style={{ fontSize: 19 }}>{`${hourNumber + 1}. `}</Text>
-        <Text h4 h4Style={{ fontSize: 19 }}>
+      <View style={{ flexDirection: "row", justifyContent: "flex-start" }}>
+        <Text h="4" style={{ fontSize: 19 }}>{`${hourNumber + 1}. `}</Text>
+        <Text h="4" style={{ fontSize: 19 }}>
           {displayedName}
         </Text>
-      </ListItem.Content>
-      {clickable && <ListItem.Chevron tvParallaxProperties={undefined} />}
-    </ListItem>
+      </View>
+      {/* {clickable && <ListItem.Chevron tvParallaxProperties={undefined} />} */}
+    </Button>
   );
 };
 
 const HoursListScreen = () => {
-  const firestoreHours = useAppSelector((state) => state.appConfig.marathonHours);
-  const [firestoreHoursWithKeys, setFirestoreHoursWithKeys] = useState<FirestoreHourWithKey[]>([]);
-  const countdown = useAppSelector((state) => state.appConfig.countdown);
+  const firestoreHours = useAppSelector((state) => state.marathon.marathonHours);
+  const [ firestoreHoursWithKeys, setFirestoreHoursWithKeys ] = useState<FirestoreHourWithKey[]>([]);
   const isConfigLoaded = useAppSelector((state) => state.appConfig.isConfigLoaded);
   const currentDate = useCurrentDate();
-  const [marathonHour, setMarathonHour] = useState(-1);
+  const [ marathonHour, setMarathonHour ] = useState(-1);
   const { width: screenWidth } = useWindowDimensions();
   const [mapOfMemorial] = useCachedFiles([
     {
@@ -137,28 +134,28 @@ const HoursListScreen = () => {
     // First programmed hour is 8:00pm or 20:00
     // Marathon is March 5th and 6th, I am hardcoding this, hope that causes no issues
     // This will set the hour to a negative number if the marathon has yet to start and should be between 0 and 23 for the duration of the marathon
-    const tempMarathonHour = 23 - differenceInHours(new Date(2022, 2, 6, 20, 0, 0, 0), currentDate);
+    const tempMarathonHour = 23 - Interval.fromDateTimes(DateTime.fromObject({ year: 2022, month: 2, day: 6, hour: 20 }), DateTime.fromJSDate(currentDate)).toDuration().as("hours");
     setMarathonHour(tempMarathonHour > 23 ? 23 : tempMarathonHour);
   }, [currentDate]);
 
   return (
     <View style={{ flex: 1 }}>
-      {marathonHour < 0 && isConfigLoaded && countdown && (
+      {/* marathonHour < 0 && isConfigLoaded && countdown && (
         <>
           <View style={{ flexGrow: 1, flexShrink: 0 }}>
             <CountdownView />
           </View>
           <View style={{ flexGrow: 2, height: 0 }} />
         </>
-      )}
+      )*/}
       {marathonHour >= 0 && (
         <>
           {mapOfMemorial && (
             <>
               <Divider width={2} />
               <Image
-                style={{ width: screenWidth, height: screenWidth * (1194 / 1598) }}
                 source={{ uri: `data:image/png;base64,${mapOfMemorial}` }}
+                width={screenWidth} height={screenWidth * (1194 / 1598)}
               />
               <Divider width={2} />
             </>
@@ -184,7 +181,7 @@ const HoursListScreen = () => {
             refreshing={!isConfigLoaded}
             onRefresh={() => {
               store.dispatch(appConfigSlice.actions.resetConfig());
-              store.dispatch(updateConfig());
+              void store.dispatch(updateConfig());
             }}
           />
         </>
@@ -192,16 +189,5 @@ const HoursListScreen = () => {
     </View>
   );
 };
-
-const style = StyleSheet.create({
-  hourRow: {
-    alignItems: "center",
-    backgroundColor: globalColors.white,
-    flexDirection: "row",
-    flex: 1,
-    height: 30,
-    justifyContent: "space-between",
-  },
-});
 
 export default HoursListScreen;
