@@ -1,87 +1,84 @@
-import { useState } from "react";
-import { View, ActivityIndicator, TextInput, ImageSourcePropType } from "react-native";
-import { Text, Button, Image } from "react-native-elements";
-import * as Linking from "expo-linking";
-import { useAssets } from "expo-asset";
+import { useNetInfo } from "@react-native-community/netinfo";
 import { nativeApplicationVersion } from "expo-application";
-import { useAppSelector } from "../../common/CustomHooks";
-import SingleSignOn from "../../common/SingleSignOn";
-import { globalStyles, globalColors, globalTextStyles } from "../../theme";
+import * as Linking from "expo-linking";
+import { Button, Image, Text } from "native-base";
+import { useState } from "react";
+import { ActivityIndicator, TextInput, View } from "react-native";
 
-import store from "../../redux/store";
-import { authSlice, logout } from "../../redux/authSlice";
-import { appConfigSlice } from "../../redux/appConfigSlice";
+import avatar from "../../../assets/avatar.png";
+import { useAppSelector } from "../../common/CustomHooks";
+import { useFirebase } from "../../common/FirebaseApp";
+import { useLinkBlueLogin } from "../../common/auth";
+import { globalColors, globalStyles, globalTextStyles } from "../../theme";
 
 /**
  * Component for "Profile" screen in main navigation
  */
 const ProfileScreen = () => {
-  const userData = useAppSelector((state) => state.auth);
+  const authData = useAppSelector((state) => state.auth);
+  const userData = useAppSelector((state) => state.userData);
   const demoModeKey = useAppSelector((state) => state.appConfig.demoModeKey);
-  const isOffline = useAppSelector((state) => state.appConfig.offline);
-  const [assets, error] = useAssets(require("../../../assets/avatar.png"));
-  const [reportLongPressed, setReportLongPressed] = useState(false);
-  const [suggestLongPressed, setSuggestLongPressed] = useState(false);
+  const { isConnected } = useNetInfo();
+  const [ reportLongPressed, setReportLongPressed ] = useState(false);
+  const [ suggestLongPressed, setSuggestLongPressed ] = useState(false);
+  const {
+    fbAuth, fbFunctions
+  } = useFirebase();
+
+  const [ loading, trigger ] = useLinkBlueLogin(fbAuth, fbFunctions);
 
   return (
     <View style={{ alignItems: "center", ...globalStyles.genericView }}>
       <>
         {
-          /* Start of still loading view */ !userData.isAuthLoaded && (
+          /* Start of still loading view */ !authData.isAuthLoaded || loading && (
             <ActivityIndicator size="large" color={globalColors.dbBlue} />
           ) /* End of still loading view */
         }
         {
-          /* Start of loaded view */ userData.isAuthLoaded && (
+          /* Start of loaded view */ authData.isAuthLoaded && (
             <>
-              {assets && (
-                <Image
-                  source={assets[0] as ImageSourcePropType}
-                  style={{ height: 64, width: 64 }}
-                />
-              )}
+              <Image
+                source={avatar}
+                height={64}
+                width={64}
+              />
               {
-                /* Start of logged in view */ userData.isLoggedIn &&
-                  !userData.isAnonymous && (
-                    <>
-                      <Text style={globalStyles.genericText}>
+                /* Start of logged in view */ authData.isLoggedIn &&
+                  !authData.isAnonymous && (
+                  <>
+                    <Text style={globalStyles.genericText}>
                         You are logged in as {userData.firstName} {userData.lastName}
-                      </Text>
-                      <Text style={globalTextStyles.italicText}>{userData.email}</Text>
-                      <Button
-                        style={{ margin: 10, alignSelf: "center" }}
-                        onPress={() => store.dispatch(logout())}
-                        title="Log out"
-                      />
-                    </>
-                  ) /* End of logged in view */
+                    </Text>
+                    <Text style={globalTextStyles.italicText}>{userData.email}</Text>
+                    <Button
+                      style={{ margin: 10, alignSelf: "center" }}
+                      onPress={() => void fbAuth.signOut()}
+                      title="Log out"
+                    />
+                  </>
+                ) /* End of logged in view */
               }
               {
-                /* Start of logged in anonymously view */ userData.isLoggedIn &&
-                  userData.isAnonymous && (
-                    <>
-                      <Text>You are logged in {isOffline ? "offline" : "anonymously"}</Text>
-                      <Button
-                        style={{ margin: 10, alignSelf: "center" }}
-                        onPress={() => {
-                          const sso = new SingleSignOn();
-                          sso.authenticate("saml-sign-in");
-                        }}
-                        title="Log in"
-                      />
-                    </>
-                  ) /* End of logged in anonymously view */
+                /* Start of logged in anonymously view */ authData.isLoggedIn &&
+                authData.isAnonymous && (
+                  <>
+                    <Text>You are logged in {isConnected ? "anonymously" : "offline"}</Text>
+                    <Button
+                      style={{ margin: 10, alignSelf: "center" }}
+                      onPress={() => trigger()}
+                      title="Log in"
+                    />
+                  </>
+                ) /* End of logged in anonymously view */
               }
               {
-                /* Start of logged out view */ !userData.isLoggedIn && (
+                /* Start of logged out view */ !authData.isLoggedIn && (
                   <>
                     <Text>You are logged out, to log in:</Text>
                     <Button
                       style={{ margin: 10, alignSelf: "center" }}
-                      onPress={() => {
-                        const sso = new SingleSignOn();
-                        sso.authenticate("saml-sign-in");
-                      }}
+                      onPress={() => trigger()}
                       title="Log in"
                     />
                   </>
@@ -94,8 +91,7 @@ const ProfileScreen = () => {
                   secureTextEntry
                   onSubmitEditing={(event) => {
                     if (event.nativeEvent.text === demoModeKey) {
-                      store.dispatch(appConfigSlice.actions.enterDemoMode());
-                      store.dispatch(authSlice.actions.enterDemoMode());
+                      // store.dispatch(authSlice.actions.enterDemoMode());
                       setReportLongPressed(false);
                       setSuggestLongPressed(false);
                     }
@@ -111,7 +107,7 @@ const ProfileScreen = () => {
                     alignSelf: "center",
                   }}
                   onPress={() => {
-                    Linking.openURL(
+                    void Linking.openURL(
                       "mailto:app@danceblue.org?subject=DanceBlue%20App%20Issue%20Report&body=What%20happened%3A%0A%3Ctype%20here%3E%0A%0AWhat%20I%20was%20doing%3A%0A%3Ctype%20here%3E%0A%0AOther%20information%3A%0A%3Ctype%20here%3E"
                     );
                   }}
@@ -123,7 +119,7 @@ const ProfileScreen = () => {
                 <Button
                   style={{ margin: 10, alignSelf: "center" }}
                   onPress={() => {
-                    Linking.openURL(
+                    void Linking.openURL(
                       "mailto:app@danceblue.org?subject=DanceBlue%20App%20Suggestion&body=%3Ctype%20here%3E"
                     );
                   }}
@@ -134,7 +130,7 @@ const ProfileScreen = () => {
                 />
                 <Text
                   style={{ textAlign: "center" }}
-                >{`Version: ${nativeApplicationVersion}`}</Text>
+                >{`Version: ${nativeApplicationVersion ?? ""}`}</Text>
               </View>
             </>
           ) /* End of loaded view */
