@@ -1,101 +1,82 @@
-import { useCallback, useEffect, useState } from "react";
+import firebaseFirestore from "@react-native-firebase/firestore";
+import { useCallback, useState } from "react";
 import { RefreshControl, SafeAreaView, ScrollView } from "react-native";
-import { getDocs, collection } from "firebase/firestore";
+
 import { useAppSelector } from "../../common/CustomHooks";
 import Standings from "../../common/components/Standings";
-import { firebaseFirestore } from "../../common/FirebaseApp";
-import { globalColors, globalStyles } from "../../theme";
-import { StandingType } from "../../types/StandingType";
-import { showMessage } from "../../common/AlertUtils";
+import { showMessage } from "../../common/util/AlertUtils";
+import { globalStyles } from "../../theme";
 import { FirestoreMoraleTeam, FirestoreTeam } from "../../types/FirebaseTypes";
-import store from "../../redux/store";
+import { StandingType } from "../../types/StandingType";
 
 /**
  * Wrapper for a Standings component
  */
 const ScoreBoardScreen = () => {
-  const { pointType } = useAppSelector((state) => state.appConfig.scoreboard);
-  const userTeamId = useAppSelector((state) => state.auth.teamId);
-  const moraleTeamId = useAppSelector((state) => state.auth.moraleTeamId);
-  const [standingData, setStandingData] = useState<StandingType[]>([]);
-  const [refreshing, setRefreshing] = useState<boolean>(true);
+  const { pointType } = useAppSelector((state) => state.appConfig.scoreboardMode);
+  const userTeamName = useAppSelector((state) => state.userData.team?.name);
+  // const moraleTeamName = useAppSelector((state) => state);
+  const [ standingData, setStandingData ] = useState<StandingType[]>([]);
+  const [ loading, setLoading ] = useState(true);
 
   const refresh = useCallback(() => {
-    setRefreshing(true);
-    let shouldUpdateState = true;
-    if (store.getState().appConfig.offline) {
-      showMessage("You seem to be offline, connect to the internet to update the scoreboard");
-    } else {
-      switch (pointType) {
-        case "spirit":
-          getDocs(collection(firebaseFirestore, "teams")).then((querySnapshot) => {
-            const tempStandingData: StandingType[] = [];
-            querySnapshot.forEach((document) => {
-              const teamData = document.data() as FirestoreTeam;
-              // Ensure we don't show the test team on the scoreboard
-              if (teamData.spiritSpreadsheetId && document.id !== "jR29Y3wJ59evnRaWWKC4") {
-                tempStandingData.push({
-                  id: document.id,
-                  name: teamData.name,
-                  points: teamData.totalSpiritPoints || 0,
-                  highlighted: userTeamId === document.id,
-                });
-              }
-            });
-            if (shouldUpdateState) {
-              setStandingData(tempStandingData);
-              setRefreshing(false);
+    switch (pointType) {
+    case "spirit":
+      void firebaseFirestore().collection("teams").get()
+        .then((querySnapshot) => {
+          const tempStandingData: StandingType[] = [];
+          querySnapshot.forEach((document) => {
+            const teamData = document.data() as FirestoreTeam;
+            // Ensure we don't show the test team on the scoreboard
+            if (teamData.spiritSpreadsheetId && document.id !== "jR29Y3wJ59evnRaWWKC4") {
+              tempStandingData.push({
+                id: document.id,
+                name: teamData.name,
+                points: teamData.totalSpiritPoints ?? 0,
+                highlighted: userTeamName === teamData.name,
+              });
             }
           });
-          break;
+          setStandingData(tempStandingData);
+        });
+      break;
 
-        case "morale":
-          getDocs(collection(firebaseFirestore, "marathon", "2022/morale-teams")).then(
-            (querySnapshot) => {
-              const tempStandingData: StandingType[] = [];
-              querySnapshot.forEach((document) => {
-                const teamData = document.data() as FirestoreMoraleTeam;
-                tempStandingData.push({
-                  id: document.id,
-                  name: `Team #${teamData.teamNumber}:\n${teamData.leaders}`,
-                  points: teamData.points,
-                  highlighted: moraleTeamId === teamData.teamNumber,
-                });
+    case "morale":
+      void firebaseFirestore().collection("marathon/2022/morale-teams").get()
+        .then(
+          (querySnapshot) => {
+            const tempStandingData: StandingType[] = [];
+            querySnapshot.forEach((document) => {
+              const teamData = document.data() as FirestoreMoraleTeam;
+              tempStandingData.push({
+                id: document.id,
+                name: `Team #${teamData.teamNumber}:\n${teamData.leaders}`,
+                points: teamData.points,
+                highlighted: false // moraleTeamId === teamData.teamNumber,
               });
-              if (shouldUpdateState) {
-                setStandingData(tempStandingData);
-                setRefreshing(false);
-              }
-            }
-          );
-          break;
-
-        case undefined:
-          if (shouldUpdateState) {
-            setStandingData([]);
-            setRefreshing(false);
+            });
+            setStandingData(tempStandingData);
           }
-          break;
+        );
+      break;
 
-        default:
-          showMessage("Failed to load valid point type configuration");
-          break;
-      }
+    case undefined:
+      setStandingData([]);
+      break;
+
+    default:
+      showMessage("Failed to load valid point type configuration");
+      break;
     }
     return () => {
-      shouldUpdateState = false;
-      setRefreshing(false);
+      setLoading(false);
     };
-  }, [pointType, moraleTeamId, userTeamId]);
-
-  useEffect(() => {
-    refresh();
-  }, [refresh]);
+  }, [ pointType, userTeamName ]);
 
   return (
     <ScrollView
       showsVerticalScrollIndicator
-      refreshControl={<RefreshControl refreshing={refreshing} onRefresh={refresh} />}
+      refreshControl={<RefreshControl refreshing={loading} onRefresh={refresh} />}
     >
       <SafeAreaView style={globalStyles.genericView}>
         <Standings titleText="Spirit Point Standings" standingData={standingData} startExpanded />
